requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic[email]
passlib
bcrypt
python-jose[cryptography]
python-dotenv
